little fix to recruiter
