--- src/dto/add-recruiter.dto.ts
@@ -1,7 +1,0 @@
-export class AddRecruiterDto {
-  id?: string;
-  email: string;
-  fullName: string;
-  company: string;
-  maxReservedStudents: number;
-}
